feat: add file operation dropdown
