+++ apps/web/components/fileManagement/fileOperationDropDown.tsx
@@ -1,0 +1,125 @@
+"use client";
+import React, { Dispatch, SetStateAction } from "react";
+import { Button } from "@/components/ui/button";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuGroup,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu";
+
+import { EllipsisVertical } from "lucide-react";
+
+import { deleteCID } from "@/services/dataVaultService";
+
+import { ipfsDelete, ipfsDownload, ipfsGetFiles } from "@/services/ipfsService";
+import { decryptWithKey } from "@/utils/encryption";
+
+// import { useWeb3AuthDisconnect } from "@web3auth/modal/react";
+import { MedicalData } from "@/interfaces/medicalData";
+const FileOperationDropDown = ({
+  walletAddress,
+  aesKey,
+  setMedicalData,
+  data,
+}: {
+  walletAddress: `0x${string}` | undefined;
+  aesKey: string | null;
+  setMedicalData: Dispatch<SetStateAction<MedicalData[]>>;
+  data: MedicalData;
+}) => {
+  //   const { disconnect } = useWeb3AuthDisconnect();
+
+  const displayContent = async (cid: string) => {
+    const decrypted = await getFileContent(cid);
+    if (!decrypted) return;
+
+    const blob = new Blob([decrypted], { type: "text/plain" });
+    const url = URL.createObjectURL(blob);
+    window.open(url, "_blank");
+  };
+  const getFileContent = async (cid: string) => {
+    if (!aesKey) return;
+    try {
+      const decryptedCid = decryptWithKey(cid, aesKey);
+      const content = await ipfsDownload(decryptedCid);
+      const decrypted = decryptWithKey(content, aesKey);
+      return decrypted;
+    } catch (error) {
+      console.error("Failed to fetch IPFS content:", error);
+      alert("Failed to load content.");
+    }
+  };
+
+  const downloadContent = async (cid: string) => {
+    const decrypted = await getFileContent(cid);
+    if (!decrypted) return;
+
+    const blob = new Blob([decrypted], { type: "text/plain" });
+    const url = URL.createObjectURL(blob);
+    const link = document.createElement("a");
+    link.href = url;
+    link.download = `medical_data_${cid}.json`;
+    document.body.appendChild(link);
+    link.click();
+    document.body.removeChild(link);
+    URL.revokeObjectURL(url);
+  };
+
+  const deleteContent = async (cid: string) => {
+    if (!aesKey) return;
+    try {
+      const decryptedCid = decryptWithKey(cid, aesKey);
+      const files = await ipfsGetFiles();
+      const uid = files.fileList.find((file) => file.cid === decryptedCid)?.id;
+      if (!uid) {
+        alert("File not found on IPFS.");
+        return;
+      }
+      await Promise.all([ipfsDelete(uid), deleteCID(walletAddress!, cid)]);
+      alert("File deleted successfully.");
+    } catch (error) {
+      console.error("Failed to fetch IPFS content:", error);
+      alert("Failed to load content.");
+    }
+  };
+  return (
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button variant="outline">
+          <EllipsisVertical />
+        </Button>
+      </DropdownMenuTrigger>
+      <DropdownMenuContent className="w-56" align="end">
+        <DropdownMenuLabel className="font-bold">File Actions</DropdownMenuLabel>
+        <DropdownMenuGroup>
+          <DropdownMenuItem onClick={() => displayContent(data.encryptedCid)} disabled={!aesKey}>
+            View Content
+          </DropdownMenuItem>
+          <DropdownMenuItem onClick={() => downloadContent(data.encryptedCid)} disabled={!aesKey}>
+            Download Content
+          </DropdownMenuItem>
+          <DropdownMenuSeparator />
+          <DropdownMenuItem
+            variant="destructive"
+            onClick={async () => {
+              if (!walletAddress) return;
+              await deleteContent(data.encryptedCid);
+              setMedicalData((prev) =>
+                prev.filter((item) => item.encryptedCid !== data.encryptedCid)
+              );
+            }}
+            disabled={!aesKey}
+          >
+            Delete Content
+          </DropdownMenuItem>
+        </DropdownMenuGroup>
+      </DropdownMenuContent>
+    </DropdownMenu>
+  );
+};
+
+export default FileOperationDropDown;
